backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
plotly==5.17.0
numpy==1.24.3
